PointsToGraph: add insertDerefPointee
This function takes the same arguments as insert, but do dereference
of location. That is, instead of adding 'p points to location' add
that 'p points to everywhere where can location point to'.

I also had to add few helper methods (well, didn't have to, but the
code is more lucid)